refactor: mapper 제거 (#313)
